back to uint

+++ com_prop_board.h
@@ -63,10 +63,10 @@ typedef struct __attribute__((__packed__)) {
     uint16_t pressure_line_N2O = 0;  // [0.0;50.0]
     uint16_t chamber_pressure = 0;  // [0.0;50.0]
 
-    float N2O_main_valve_position = 0;   // [0.0; 100.0]
-    float ETH_main_valve_position = 0;   // [0.0; 100.0]
-    float N2O_main_valve_current = 0; // current as in Amps
-    float ETH_main_valve_current = 0;
+    uint16_t N2O_main_valve_position = 0;   // [0.0; 100.0]
+    uint16_t ETH_main_valve_position = 0;   // [0.0; 100.0]
+    uint16_t N2O_main_valve_current = 0; // current as in Amps
+    uint16_t ETH_main_valve_current = 0;
     uint16_t gimbal_x_position = 0;    // [0.0; 100.0]
     uint16_t gimbal_y_position = 0;    // [0.0; 100.0]
